Windows changes to reduce #s of warnings in Windows builds. Alignment sequence identifiers more consistently intptr_t rather than int or size_t (can't be size_t because that is unsigned, and -1 is often used as a "magic" sequence identifier). 1. Alignment::getSeqID() now returns intptr_t, rather than int 2. Alignment::getNSeq() now returns intptr_t, rather than size_t 3. Declared StringInt64Map (in statespace.h) 4. Used that, in PhyloTree::computeSiteConcordance(), in place     of StringIntMap. 5. Declared Int64Vector (in utils/vectortypes.h) 6. In whtest .c and .h files, nr_basen and taxa are now inptr_t, not int.

+++ alignment/alignment.h
@@ -371,7 +371,7 @@ public:
     /**
             @return number of sequences
      */
-    inline size_t getNSeq() const {
+    inline intptr_t getNSeq() const {
         return seq_names.size();
     }
 
@@ -451,7 +451,7 @@ public:
             @param seq_name sequence name
             @return corresponding ID, -1 if not found
      */
-    int getSeqID(const string &seq_name) const;
+    intptr_t getSeqID(const string &seq_name) const;
 
     /**
             @return length of the longest sequence name
